+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - It is now possible to `await` a `DOMQuery.remove`. Note that this changes
   the return value of `DOMQuery.remove`, which uses to return `self`.
   https://github.com/Textualize/textual/issues/1094
+- Added Pilot.wait_for_animation
 
 ### Changed